cpu/stm32f0: even more factorize
